Update Calculator
First off, I wrapped everything up in neat little functions to keep things that would be able to cause bugs from causing bugs; shit only runs if it works now.

Second, I limited the accepted responses to $choice in favor of things not throwing a bitch fit if the user says the wrong thing;

Third, I made sure it returns exit() if we actually say no

Fourth, I got rid of floats unless absolutely neccessary; not even every division operation uses a float in this.
We first check to see if there's a remainder left over from the division at all, and if it's not a comfy fit THEN we use float.

Fifth, I settled on a nomenclature and stuck with it; lotta tutorials jump all over the place with variable names Like_this OrThis orThis and it's not helpful to jump around.  Totally up to taste; I prefer to use two-word names with the first letter capitalized and no underscore LikeThis() Simple, and elegant. But again, taste. Just stick whith whatever style suits you.